backend/requirements.txt
fastapi
uvicorn
langgraph
langchain-groq
ddgs
pydantic-settings
python-dotenv
beautifulsoup4
requests
